+++ packages/storybook/context/i18n.context.js
@@ -2,17 +2,22 @@ import Vue from 'vue';
 import VueI18n from 'vue-i18n';
 
 import { addDecorator } from '@storybook/vue';
-import { ENGLISH_LOCALE } from '../constants/globalisation';
 
-Vue.use(VueI18n);
+export default () => {
+    Vue.use(VueI18n);
 
-const i18n = new VueI18n({
-    locale: ENGLISH_LOCALE,
-    fallbackLocale: ENGLISH_LOCALE,
-    messages: {}
-});
+    const ENGLISH_LOCALE = 'en-GB';
 
-addDecorator(() => ({
-    template: '<story/>',
-    i18n
-}));
+    const i18n = new VueI18n({
+        locale: ENGLISH_LOCALE,
+        fallbackLocale: ENGLISH_LOCALE,
+        messages: {}
+    });
+
+    addDecorator(() => ({
+        template: '<story/>',
+        i18n
+    }));
+
+    return i18n;
+};