+++ test/utils/webdriverio-extensions.js
@@ -1,0 +1,3 @@
+module.exports = browser => ({
+    doesElementExist: selector => $$(selector).length > 0
+});